datapath/fake: Move NodeHandler to fake/types package
This is the first commit in a series of commits which moves various mock
types from `pkg/datapath/fake` to `pkg/datapath/fake/types`. The
motivation behind this change lies in the fact that `pkg/datapath/fake`
serves two purposes:

1. It contains a "fake-datapath" hive cell providing various (mock)
   dependencies needed to run the `test/controlplane` tests. As such,
   the `cells.go` file defining the "fake-datapath" cell imports a lot of
   other Go packages.
2. It contains a few mock implementation of various datapath-related
   bookkeeping structures, such as `NodeHandler`, `MTU` or
   `NodeAddressing`. Those are used by the "fake-datapath" cell, but also
   commonly used by unit tests in various packages.

This means that any package referenced by the "fake-datapath" cell (and
any transitively reachable packages) cannot use any of the mock types in
unit testing, as we would otherwise introduce a circular dependency.
This means blocks the "fake-datapath" cell from providing certain types
provided by the real datapath cell (found in `pkg/datapath/cells.go`).

Therefore, this series of commits attempts to separate the two use-cases
by splitting the `pkg/datapath/fake` package into two.

The existing package will continue to provide the "fake-datapath" cell
(in `pkg/datapath/fake/cells.go`), mirroring the real
`datapath/cells.go`, where as a new `pkg/datapath/fake/types` package
will provide mock types referenced in unit tests by various other
packages, in some sense mirroring the real `pkg/datapath/types`.

In this first commit, we move the fake `NodeHandler` implementation.
